+++ changelog.md
@@ -1,5 +1,7 @@
 # Changelog
 
+Go to the [dev branch](https://github.com/aquarial/discord-haskell/blob/dev/changelog.md) for changes not merged into master yet
+
 ### dev
 
 Add requests: ModifyChanPositions